update tweet.ts written()
Got rid of the hashtag and the links then made it so that the tweets are lowercase. made an array of the words that automated tweets have and have each tweet go through that and replace with '' so that i can get the length of the new tweet. then if the tweet is less than 3 then it is a written tweet else it contains user tweets