Update descriptions.js
Get the tweet array with the runkeeper, then the making it so that it is only including the written tweets.  And adding to event handler, make the ids from html, and then making the event listeners to clear the table and other data. additionally, using the tweet array that only has written tweets to match and then use it to update table.